+++ example/read.dart
@@ -5,9 +5,9 @@ import 'package:rpi_gpio/wiringpi_gpio.dart';
 /// Read current values for pins 0 - 7
 main() async {
 
-  // Initialize the hardware
+  // Initialize the GPIO API
   // See read_with_mocks.dart for testing on non-RaspberryPi platforms
-  Pin.hardware = new WiringPiGPIO();
+  Pin.gpio = new WiringPiGPIO();
 
   for (int pinNum = 0; pinNum < 8; ++pinNum) {
     var p = pin(pinNum, Mode.input);